requirements.txt
python-dotenv==0.20.0
# pytube==12.1.0
git+https://github.com/pishiko/pytube.git@master#egg=pytube
requests==2.27.1
tweepy==4.14.0